use wayfair SKU when item mapping mode is SKU (#164)
* prefer the wayfair SKU when getting product IDs in SKU mode
* fallback to using the first SKU when getting product IDs in SKU mode and no Wayfair one is found.
* fix tests
* remove casts of referrer ID (float) to int.

+++ src/Core/Helpers/AbstractConfigHelper.php
@@ -44,9 +44,9 @@ abstract class AbstractConfigHelper {
   /**
    * Returns the PlentyMarkets identifier for Wayfair's Order Referrer value
    *
-   * @return int
+   * @return float
    */
-  abstract public function getOrderReferrerValue(): int;
+  abstract public function getOrderReferrerValue(): float;
 
   /**
    * Get the value of the user-configured Stock Buffer setting
